Consolidate KeycloakManager to single file in api/utils/keycloak_utils.py
