+++ authorization/backends.py
@@ -4,7 +4,7 @@ from django.contrib.auth import get_user_model
 from django.contrib.auth.backends import ModelBackend
 from django.http import HttpRequest
 
-from authorization.keycloak import keycloak_manager
+from api.utils.keycloak_utils import keycloak_manager
 from authorization.models import User
 
 
@@ -14,10 +14,7 @@ class KeycloakAuthenticationBackend(ModelBackend):
     """
 
     def authenticate(  # type: ignore[override]
-        self,
-        request: Optional[HttpRequest] = None,
-        token: Optional[str] = None,
-        **kwargs: Any
+        self, request: Optional[HttpRequest] = None, token: Optional[str] = None, **kwargs: Any
     ) -> Optional[User]:
         """
         Authenticate a user based on a Keycloak token.
@@ -44,9 +41,7 @@ class KeycloakAuthenticationBackend(ModelBackend):
 
         # Get user roles and organizations from the token
         roles: List[str] = keycloak_manager.get_user_roles(token)
-        organizations: List[Dict[str, Any]] = keycloak_manager.get_user_organizations(
-            token
-        )
+        organizations: List[Dict[str, Any]] = keycloak_manager.get_user_organizations(token)
 
         # Sync the user information with our database
         user: Optional[User] = keycloak_manager.sync_user_from_keycloak(
